[IMPROVEMENT] Refactor Profile page and use address subscriptions to automatically refresh address details on the page (#131)
Ref: DGDG-300

This refactor modularizes the Profile page by separating its parts into the following components: `<ProfileUserInfo/>`, `<UserAddressStats/>` (see PR #130), `<ProfileActivitySummary/>`, and `<ModeratorRequirements/>`. This should make it easier to maintain and to see which data is relevant for each section.

All components now use GraphQL queries and subscriptions to check for changes in the user data, address details, and KYC data.

**Test Plan**

The following should automatically change on the page when the server data changes:
- Username (via set username)
- Email (via set email. Test by setting the email from the KYC button and the Set Email button)
- User status (if they are a participant, moderator, etc.)
- KYC status (for all status changes except for when the KYC expires)
- Reputation and Quarter points (test with redeem badge)
- Locked DGD and locked stake (test by locking DGD both from the header and the moderator requirements section)
- Remaining stake to lock to get moderator status (via lock DGD)

Regression tests for the Profile page should pass.
